Backend Primordial Phase
This is an extremely early version of some backend concepts. Do not
consider this remotely near the finished product. Load the
backEndTestCase.php to test out some array and for loop methods as well
as taking in user input and feeding it to the php side of things. If
it's not clear than check this out:

https://www.youtube.com/watch?v=dQw4w9WgXcQ