"Create Poll" Pages
They aren't pretty, but here's a Poll Creation tool to test out
inserting polls into the database. Open the createPoll.php page to get
user input, click submit, and the createPollBackend.php will concatenate
the user input into a string and attempt to insert it into the database.

When inserting there is currently an error with the Poll ID